requirements.txt
# AI_OS — Python dependencies
# Восстановить: pip install -r requirements.txt

# AI APIs
openai==2.24.0
anthropic

# Web framework (для WB_BOT web/)
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1
python-multipart==0.0.22
jinja2==3.1.6

# HTTP
httpx==0.28.1
requests==2.33.1

# Data validation
pydantic==2.12.5

# Utilities
python-dotenv==1.2.1
tqdm==4.67.3
tenacity==9.1.4

# Dev / testing
pytest==9.0.2
